Create 19-Closest In BST
Closest in BST

Given a BST and an integer. Find the least absolute difference between any node value of the BST and the given integer.

(https://www.geeksforgeeks.org/problems/find-the-closest-element-in-bst/1)

Example 1:

Input:
        10
      /   \
     2    11
   /  \ 
  1    5
      /  \
     3    6
      \
       4
K = 13
Output: 2
Explanation: K=13. The node that has value nearest to K is 11. so the answer is 2
Example 2:

Input:
      8
    /   \
   1     9
    \     \
     4    10
    /
   3
K = 9
Output: 0
Explanation: K=9. The node that has value nearest to K is 9. so the answer is 0.
Constraints:
1 <= Number of nodes <= 10^5
1 <= Value stored at nodes(data), K <= 10^5